[Feat] Divide Api interface
